deployment/frontend/requirements.txt
requests
streamlit
opencv-python-headless
numpy
Pillow
typing_extensions